Extract static TUV methods to new util class
